Extract Method
Extracted 3 methods to modify the quality depending of the object type.

Next step could be "Replace Type code with State/Strategy" but... the
Kata has a constrain:

"Feel free to make any changes to the UpdateQuality method and add any
new code as long as everything still works correctly. However, do not
alter the Item class or Items property as those belong to the goblin in
the corner who will insta-rage and one-shot you as he doesn't believe in
shared code ownership (you can make the UpdateQuality method and Items
property static if you like, we'll cover for you)."